+++ src/adstractai/constants.py
@@ -8,7 +8,7 @@ AD_ACK_ENDPOINT = "/api/ad-injection/acknowledge/"
 
 # Versioning and identification
 SDK_NAME = "adstractai-python"
-SDK_VERSION = "1.0.0"
+SDK_VERSION = "1.0.1"
 TYPE = "sdk"
 
 DEFAULT_TIMEOUT_SECONDS = 100